fix(weekly-digest): route new error issues query through error_tracking facade
error_tracking is an isolated product; importing ErrorTrackingIssue directly
broke the tach interface check. Move the query into the facade and register
the new error-issue and usage-trends lookup activities in the workflow test.
